refactor: move mypage components to features directory - PIT-273
